[C8] Split engine.go by responsibility (#2234)
model/feature/engine.go was 536 LOC, over the 500-line ceiling. Split the
feature recompute engine along its responsibilities; each file is now under
the ceiling:

- engine.go (218): the PartFeatures collection + configuration (relief/corner/
  bend/miter/scale/font resolvers), Add/Remove and the unique-naming + lookup
  (ByID/ByName/UniqueName/Count/Item/Result).
- engine_recompute.go (264): the recompute loop and health/quarantine —
  Recompute/evaluateFrom/evaluate/evaluateBody, safeRecompute (panic capture),
  the source tool/delta derivation and classify (Sick reporting).
- engine_graph.go (82): the feature graph/ordering — dirty tracking
  (MarkDirty/RequiresUpdate/MarkAllDirty) and the ordering queries
  (effectiveEnd/earliestDirty/prefixBodies/sickBefore/dependsOnSick).

Behaviour-preserving: split programmatically by receiver-qualified
declaration, verified every non-comment code line is byte-identical (only
import redistribution + one gofmt one-liner alignment differ, the latter
restored); imports resolved per file with goimports. Feature package builds,
archguard passes, lint 0. Net delta of recognizers / tolerance constants /
fallback sites / type assertions: 0.

Part of #2183 (C8: structure/size limits). Closes #2234.

(cherry picked from commit a495b348c1f9664758ab2eb16f7e92887847bb55)
